move classes to proper domain for compliance
