+++ src/test/java/ch/heigvd/res/chill/domain/capito27/RedStoutTest.java
@@ -1,6 +1,7 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.capito27;
 
 import ch.heigvd.res.chill.domain.Bartender;
+import ch.heigvd.res.chill.domain.capito27.RedStout;
 import ch.heigvd.res.chill.protocol.OrderRequest;
 import ch.heigvd.res.chill.protocol.OrderResponse;
 import org.junit.jupiter.api.Test;
@@ -21,7 +22,7 @@ class RedStoutTest {
     @Test
     void aBartenderShouldAcceptAnOrderForPunkIPA() {
         Bartender jane = new Bartender();
-        String productName = "ch.heigvd.res.chill.domain.wasadigi.RedStout";
+        String productName = "ch.heigvd.res.chill.domain.capito27.RedStout";
         OrderRequest request = new OrderRequest(3, productName);
         OrderResponse response = jane.order(request);
         BigDecimal expectedTotalPrice = RedStout.PRICE.multiply(new BigDecimal(3));
